+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ### 0.0.1
   - Initial release
+  - Algorand Smart Contract Project type
+  - Basic Code Completion and Syntax Highlighter support for TEAL